remove tooltip, because we use popovers
there are some problems if we use tooltips and popovers in same code -
the settings for popover blocks the tooltip code.

Popovers require Tooltips to be included. Aside from the visual
difference, popovers have an option to display both a title and content,
while tooltips only have an option to display a title.