New form type 'participationForm'
- type navbar doesn´t fit the html for the new modal.  So we need the
default type, but this has the disadvantage to have the "post" action
and we need the "get" for all other stuff. So this is an interim and we
should figure out how to do this in future.
I would prefer definitly using $_POST for the client form in regards of
security