Issue 1067: Implement admin notification mails for all user/profile/membership changes
Notifications for all changes to a user, their profile fields and memberships are implemented via a dedicated class ChangeNotification, which queues all changes and sends out the mail notifications (if system notifications are enabled in the configuration) when the PHP script is done processing the whole page (but just before all global objects are destroyed).
A global object gChangeNotifications is automatically created and all changes in the User and TableMembers classes are properly logged to this object. If any plugin or code directly modifies the database, it needs to use any of the member functions of the ChangeNotification class to register the change.

In the past, all profile field changes were logged to the database already. This functionality has now been moved to the ChangeNotification class, so this class is the control center for all change notifications and logging. Implementing history log entries for user creation/deletion, password/username changes or membership changes can be easily added in the corresponding functions in the ChangeNotification class.

Fixes #1067