In all lists, make the running number a link to the user in html mode
Currently, only first and last name columns in html lists were links to the user. If a list did not contain either of these columns (e.g. because it displayed the company name for corporate members), no link to edit the member was possible.

The first column of the list is always a running number, so this patch also adds the link to the line number, making sure there is always a way to got the member and edit it.