User fields: If a regexp contains a forward slash, it needs to be escaped, because we manualla prepend/append a forward slash for preg_match. That's opaque to the user, so requiring them to escape slashes is not the right approach.
Fig regexp handling: Joining a "/" as delimiter with a user-provided string to insert into a preg_match call just screams for troubles. -> Find a proper delimiter that is not contained in the regexp.

So far, a preg_match call was done by joining '/' with user-entered text and using that as delimited regexp. The user-entered text was never sanitized and no escaping was done, so a slash in the regexp cause the regexp patter to terminate early.
At least, I'm not aware of an security attack vector, but it can crash the execution if the regexp contains the delimiter.
As a solution, I check a number of seldomly-used characters to serve as delimiters for the regexp, until I find a delimiter that is not contained in the regexp.