General DB logging: Add a "View changelog" button wherever possible
Since we log basically all changes to objects stored in the database (as long as they use the generic TableAccess base class to write data to the databse), we can show a filtered changelog for each type of object (and even for each individual object). This patch simply adds the corresponding links. The core functionality (writing the change log entries to the DB, as well as providing a nice changelog page, which takes URL params to filter by DB table and object ID / UUID) is added in previous commits.