Don't disable all exceptions in the ChangeNotification constructor
This is a global setting that is not expected to happen. In particular, it prevents inserting new preference entries into the database, because that relies on an exception when the prf_name cannot be found. With exceptions disabled, the preferences managet will always try to update and never insert a new setting.