Changelog: Convert role dependency handling to Entity-based DB access rather than using direct SQL statements -> logging of role dependencies also implemented.
Unfortunately, the adm_role_dependencies table is the only table left that does not use an auto-increment id, but employs a two-column key (parent-child), which in some parts collides with the assumptions in the Entity class (based on a single keyColumnName).

Also, there is now the RoleDependency value object and the RoleDependencies subclass of the Entity class (for the actual DB record). They should probably be consolidated into one class, derived from Entity. It might even be a better idea to add an auto-increment column to the adm_role_dependencies class, so the DB handling is in line with all other tables.