Changelog: Fix changelog filtered by user UUID; Show notice if changelog is filtered due to missing permissions (own user data is visible, other's not).
For the notice to be shown, I extended the DataTable json with a 'notice' array, where each entry "ID: text" means that the <div> with id #ID will be set to the given context and shown. If text is empty, the div will be hidden.

This is a generic extension in our HtmlTable class with dynamic / server-side content loading.