Generalize RolesRightsData to use Entity sub-classes
Currently, RolesRightesData will create an Entity object for all classes that are not explicitly listed. In the future, more classes might get roles rights support. If the dedicated Entity-sub-class is not used, the changelog implementation for that table is not applied.

On the other hand, ChangelogService::getObjectForTable already provides the functionality to create an (empty) object for a given DB table... Using that method solves the problem.