Login redirect: Allow redirecting to a POST URL after successful login
The login form so far (optionally) stores the current URL and redirects to the URL after a successful login. However, for SSO, some clients use a POST request to the SSO endpoints, so after successful login the POST request needs to be repeated.

This is implemented via three steps:
1) Before the login form in shown, the $_POST array is stored in the session ($_SESSION['login_forward_url_post']), alongside the login_forward_url
2) When the login form is processed, the session is checked for the login_forward_url_post entry. If it exists, it adds it to the returned JSON
3) The JS code handling the login response uses either self.location.href to switch to the URL if no post arguments are given, or creates a hidden form, adds the post variables as hidden inputs and submits the form, resulting in the proper POST request being sent to the URL.

This is fully implemented, and also used in the SAML 2.0 SSO code, as Keycloak sends the SSO request as POST (most other clients use GET).