Theming support: Add preferences for color, logo, favicon and custom css file
The most common theme modifications are colors, links to logo and favicon, as well as tiny CSS adjustments. In the past, these were hardcoded into the theme, so any change required modifying the themen (or copying it to a custom theme).

This patch adds configuration variables for the most common design adjustments, so that they no longer require a custom themen:
  * primary (header bar) and secondary (sidebar) colors
  * logo and favicon
  * custom css file included after the theme's css