Change strict equality to loose equality for hasRequiredInput method
When you go to the registration page (profile_new.php), the required fields that are not hardcoded (i.e., username, password, etc) but defined as required (i.e., email) in the profile field configuration/database (usf_required_input) are not being made required in the HTML form. This is because the operator is too strict.