DataTables Javascript error after installation when language not set yet (Fix issue #1895)
Immediately after a fresh installation, no language is set. The contacts datatable, however, tried to load the translated .json by simply inserting the iso code into the url, leading to a broken URL.

As a simple fix, default to 'en' if no language is set yet. Once a language is set, it cannot be unset manually in the settings.

Fixes #1895